Learning outcomes  of Coffee machine simulator.
The machine works with typical products: coffee, milk, sugar, and plastic cups; if it runs out of something, it shows a notification. You can get three types of coffee: 
                  1.espresso
                  2.cappuccino 
                  3. latte
Since nothing’s for free, it also collects the money.
This project allows to get a taste of Java. Practice working with classes, challenge yourself with loops and conditions, and get more confident in Java.